+++ src/main/java/io/anserini/kg/FreebaseNode.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.anserini.kg.freebase;
+package io.anserini.kg;
 
 import org.openrdf.model.Literal;
 import org.openrdf.model.ValueFactory;
